Changes to ProductsManagmentController.Java
