+++ productsManagement/src/main/java/com/mcastillo/productsManagement/controller/ProductsManagementController.java
@@ -3,11 +3,12 @@ package com.mcastillo.productsManagement.controller;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.mcastillo.Product;
 import com.mcastillo.productsManagement.service.ProductsManagementService;
-import com.mcastillo.productsManagement.service.ProductsManagementServiceInterface;
+import com.mcastillo.productsManagement.service.ProductsManagementServiceImpl;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
+import java.util.List;
 import java.util.concurrent.TimeoutException;
 
 @RestController
@@ -16,32 +17,32 @@ import java.util.concurrent.TimeoutException;
 @CrossOrigin
 public class ProductsManagementController {
 
-	private final ProductsManagementServiceInterface service;
+	private final ProductsManagementService productsManagementService;
 
-	public ProductsManagementController(ProductsManagementService service) {
-		this.service = service;
+	public ProductsManagementController(ProductsManagementServiceImpl productsManagementService) {
+		this.productsManagementService = productsManagementService;
 	}
 
 	@GetMapping("/products")
-	public ResponseEntity<?> getProducts() throws TimeoutException {
-		return ResponseEntity.ok(service.getProducts());
+	public ResponseEntity<String> getProducts() throws TimeoutException {
+		return ResponseEntity.ok(productsManagementService.getProducts());
 	}
 
 	@PostMapping("/products")
-	public ResponseEntity<?> createProduct(@RequestBody Product product) throws JsonProcessingException, TimeoutException {
-			service.createProduct(product);
-			return ResponseEntity.ok("Product created with id: " + product.getId());
+	public ResponseEntity<String> createProduct(@RequestBody Product product) throws JsonProcessingException, TimeoutException {
+		String response = productsManagementService.createProduct(product);
+		return ResponseEntity.status(HttpStatus.CREATED).body(response);
 	}
 
 	@PutMapping("/products/{id}")
-	public ResponseEntity<?> updateProduct(@PathVariable String id, @RequestBody Product product) throws JsonProcessingException, TimeoutException {
-			service.updateProduct(product);
-			return ResponseEntity.ok("Product updated with id: " + id);
+	public ResponseEntity<String> updateProduct(@PathVariable String id, @RequestBody Product product) throws JsonProcessingException, TimeoutException {
+		String response = productsManagementService.updateProduct(product);
+		return ResponseEntity.status(HttpStatus.OK).body(response);
 	}
 
 	@DeleteMapping("/products/{id}")
-	public ResponseEntity<?> deleteProduct(@PathVariable int id) throws TimeoutException, JsonProcessingException {
-			service.deleteProduct(id);
-			return ResponseEntity.ok("Product deleted with id: " + id);
+	public ResponseEntity<String> deleteProduct(@PathVariable int id) throws TimeoutException, JsonProcessingException {
+		productsManagementService.deleteProduct(id);
+		return ResponseEntity.status(HttpStatus.NO_CONTENT).body(null);
 	}
 }
